Exctractors splitted
