+++ interface/src/js/decoders/typeDecoder.js
