+++ interface/src/js/decoders/weatherDecoder.js
