+++ interface/src/js/extractors/extractPveBoss.js
@@ -1,0 +1,6 @@
+export function extractPveBoss(array) {
+    return {
+        Name: array[0], QuickMove: array[1], ChargeMove: array[2],
+        Tier: array[3], quickMovePool: "", chargeMovePool: "",
+    }
+}
